+++ testdata/de/dm/intellij/liferay/language/osgi/ReferenceWithoutComponentInspectionTest/ClassWithComponent.java
@@ -1,0 +1,10 @@
+import org.osgi.service.component.annotations.Component;
+import org.osgi.service.component.annotations.Reference;
+
+@Component
+public class ClassWithComponent {
+
+    @Reference
+    private MyObject object;
+
+}