+++ testdata/de/dm/intellij/liferay/language/osgi/ReferenceWithoutComponentInspectionTest/MyObject.java
@@ -1,0 +1,6 @@
+import org.osgi.service.component.annotations.Component;
+
+@Component
+public class MyObject {
+
+}