+++ testdata/de/dm/intellij/liferay/language/osgi/ReferenceWithoutComponentInspectionTest/org/osgi/service/component/annotations/Reference.java
@@ -1,0 +1,4 @@
+package org.osgi.service.component.annotations;
+
+public @interface Reference {
+}